requirements.txt
# Core BDD framework
behave==1.2.6

# Web UI automation
selenium==4.18.1
webdriver-manager==4.0.1

# API automation
requests==2.31.0
jsonschema==4.21.1

# Reporting
allure-behave==2.13.2
allure-python-commons==2.13.2

# Test utilities
pytest==8.0.0
assertpy==1.1

# Configuration & data handling
python-dotenv==1.0.1
PyYAML==6.0.1

# Logging & date utilities
loguru==0.7.2
python-dateutil==2.8.2

#html report
behave-html-formatter

#parallel execution

#read excel file
openpyxl




